Updated the VS project files to incude the new audio_backend files. In that process I updated nes.h so it no longer has to include the cpu, apu, ppu or cart header files, which both means reduced compile times when those files are changed and also I don't need to tell the FrontEnd project file about anything except the folder containing nes.h. Finally added a quick fix that resets the PPU's frame counter and frame start times after resuming from pause. This prevents it from telling the APU that the emulator is running at a ridiculously slow speed

+++ src/Emulator/apu.cc
@@ -746,7 +746,7 @@ void APU::SetFrameLength(int32_t length)
         else
         {
             frequency = (16667.f / length) * CpuFrequency;
-            EffectiveCpuFrequency = frequency;
+            EffectiveCpuFrequency = static_cast<int32_t>(frequency);
         }
 
         // Fraction represents the fractional part of the CPU frequency divided by the sample rate
